% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_binary.R
\name{binary_read_val}
\alias{binary_read_val}
\title{Binary protocol: read value from binary buffer}
\usage{
binary_read_val(inbuf, ttype, spec = NA, decode_response = TRUE)
}
\arguments{
\item{inbuf}{binary buffor}

\item{ttype}{type of value}

\item{spec}{specification of value}

\item{decode_response}{for string decode binary as chars}
}
\value{
value of type ttype
}
\description{
Binary protocol: read value from binary buffer
}
