% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_binary.R
\name{binary_write_val}
\alias{binary_write_val}
\title{Binary protocol: write value to binary buffer}
\usage{
binary_write_val(outbuf, ttype, val, spec = NA)
}
\arguments{
\item{outbuf}{binary buffor}

\item{ttype}{type of value}

\item{val}{value to write}

\item{spec}{specification of value}
}
\description{
Binary protocol: write value to binary buffer
}
