% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hexlify}
\alias{hexlify}
\title{hexlify}
\usage{
hexlify(byte_array, delimeter = " ")
}
\arguments{
\item{byte_array}{raw array}

\item{delimeter}{separation character}
}
\value{
string
}
\description{
String representation of raw array
}
