% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrift.R
\name{init_func_generator}
\alias{init_func_generator}
\title{init_func_generator}
\usage{
init_func_generator(cls, spec)
}
\arguments{
\item{cls}{R6 class}

\item{spec}{specification}
}
\value{
constructor function
}
\description{
Generate `initialize` function based on TPayload$default_spec
}
