% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{make_client}
\alias{make_client}
\title{Create client side thrift API}
\usage{
make_client(service, host = "localhost", port = 9090,
  proto_factory = TBinaryProtocolFactory$new(),
  trans_factory = TBufferedTransportFactory$new())
}
\arguments{
\item{service}{a}

\item{host}{a}

\item{port}{a}

\item{proto_factory}{a}

\item{trans_factory}{a}
}
\description{
Create client side thrift API
}
