% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{make_server}
\alias{make_server}
\title{Create server side thrift API}
\usage{
make_server(service, handler, host = "localhost", port = 9090,
  proto_factory = TBinaryProtocolFactory$new(),
  trans_factory = TBufferedTransportFactory$new())
}
\arguments{
\item{service}{parsed service}

\item{handler}{R6 class implementing service}

\item{host}{server host}

\item{port}{port server tcp port}

\item{proto_factory}{factory that generates protocol implementation}

\item{trans_factory}{factory that generates transport implementation}
}
\description{
Create server side thrift API
}
