% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrift.R
\name{parse_spec}
\alias{parse_spec}
\title{parse_spec}
\usage{
parse_spec(ttype, spec = NA)
}
\arguments{
\item{ttype}{type}

\item{spec}{specification}
}
\value{
string representation
}
\description{
String representation of specification
}
