% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{t_load}
\alias{t_load}
\title{Load thrift file as a R6 instance.}
\usage{
t_load(path, module_name = NA, include_dirs = NA)
}
\arguments{
\item{path}{file path to parse, should be a string ending with '.thrift'}

\item{module_name}{the name for parsed module, the default is the basename
without extension of `path`}

\item{include_dirs}{directories to find thrift files while processing
the `include` directive, by default: ['.']}
}
\value{
Thrift R6 class instance
}
\description{
The module loaded and objects inside may only be pickled if module_name
was provided.
}
