% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_binary.R
\docType{class}
\name{TBinaryProtocol}
\alias{TBinaryProtocol}
\title{protocol_binary.R}
\format{An \code{\link{R6Class}} generator object}
\usage{
TBinaryProtocol
}
\description{
Binary implementation of the Thrift protocol driver.
}
\keyword{data}

