% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_binary.R
\docType{data}
\name{TBinaryProtocolFactory}
\alias{TBinaryProtocolFactory}
\title{protocol_binary.R}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TBinaryProtocolFactory
}
\description{
protocol_binary.R
}
\keyword{datasets}

