% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_buffered.R
\docType{class}
\name{TBufferedTransport}
\alias{TBufferedTransport}
\title{transport_buffered.R
Class that wraps another transport and buffers its I/O.}
\format{An \code{\link{R6Class}} generator object}
\usage{
TBufferedTransport
}
\description{
The implementation uses a (configurable) fixed-size read buffer
but buffers all writes until a flush is performed.
}
\keyword{data}

