% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\docType{class}
\name{TBufferedTransportFactory}
\alias{TBufferedTransportFactory}
\title{TBufferedTransportFactory ...}
\format{An \code{\link{R6Class}} generator object}
\usage{
TBufferedTransportFactory
}
\description{
This class is ...
}
\keyword{datasets}
