% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrift.R
\docType{class}
\name{TClient}
\alias{TClient}
\title{TClient}
\format{An \code{\link{R6Class}} generator object}
\usage{
TClient
}
\description{
TClient implements client api of thrift service.
}
\keyword{datasets}
