% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_memory.R
\docType{class}
\name{TMemoryBuffer}
\alias{TMemoryBuffer}
\title{TMemoryBuffer}
\format{An \code{\link{R6Class}} generator object}
\usage{
TMemoryBuffer
}
\description{
Wraps a BytesIO object as a TTransport.
}
\keyword{datasets}
