% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_proper_struct}
\alias{to_proper_struct}
\title{to_proper_struct}
\usage{
to_proper_struct(thrift_spec_list, default_spec)
}
\arguments{
\item{thrift_spec_list}{raw array}

\item{default_spec}{separation character}
}
\value{
R6 class
}
\description{
Help method for tests. It changes predefined structure to parsed thrift
instead of parsing file.
}
