% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrift.R
\docType{class}
\name{TPayload}
\alias{TPayload}
\title{TPayload}
\format{An \code{\link[R6]{R6Class}} generator object}
\usage{
TPayload
}
\description{
Base class for all complex types of api.
}
\keyword{datasets}
