% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_socket.R
\docType{class}
\name{TServerSocket}
\alias{TServerSocket}
\title{TServerSocket}
\format{An \code{\link[R6]{R6Class}} generator object}
\usage{
TServerSocket
}
\description{
Socket implementation for server side.
}
\keyword{datasets}
