% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_socket.R
\docType{class}
\name{TSocket}
\alias{TSocket}
\title{transport_socket.R
Socket implementation for client side.}
\format{An \code{\link{R6Class}} generator object}
\usage{
TSocket
}
\arguments{
\item{host}{the host to connect to}

\item{port}{the (TCP) port to connect to}
}
\description{
The implementation uses a (configurable) fixed-size read buffer
but buffers all writes until a flush is performed.
}
\keyword{data}

