% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\docType{class}
\name{TTransportBase}
\alias{TTransportBase}
\title{Base class for Thrift transport layer.}
\format{An \code{\link{R6Class}} generator object}
\usage{
TTransportBase
}
\description{
Base class for Thrift transport layer.
}
\keyword{data}

