% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{make_client}
\alias{make_client}
\title{Create client side thrift API}
\usage{
make_client(service, host = "localhost", port = 9090)
}
\arguments{
\item{service}{a}

\item{host}{a}

\item{port}{a}

\item{proto_factory}{a}

\item{trans_factory}{a}

\item{timeout}{a}

\item{cafile}{a}

\item{ssl_context}{a}

\item{certfile}{a}

\item{keyfile}{a}
}
\description{
Create client side thrift API
}
