% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse}
\alias{parse}
\title{Parse a single thrift file to module object, e.g.::}
\usage{
parse()
}
\value{
Thrift module
}
\description{
>>> from thriftpy.parser.parser import parse
>>> note_thrift = parse("path/to/note.thrift")
<module 'note_thrift' (built-in)>
}

