% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\docType{class}
\name{TBinaryProtocolFactory}
\alias{TBinaryProtocolFactory}
\title{TBinaryProtocolFactory ...}
\format{An \code{\link{R6Class}} generator object}
\usage{
TBinaryProtocolFactory
}
\description{
This class is ...
}
\keyword{datasets}

