% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_buffered.R
\docType{class}
\name{TBufferedTransport}
\alias{TBufferedTransport}
\title{TBufferedTransport}
\format{An \code{\link[R6]{R6Class}} generator object}
\usage{
TBufferedTransport
}
\description{
Class that wraps another transport and buffers its I/O.
}
\keyword{datasets}
