% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_buffered.R
\docType{data}
\name{TBufferedTransportFactory}
\alias{TBufferedTransportFactory}
\title{transport_buffered.R}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TBufferedTransportFactory
}
\description{
transport_buffered.R
}
\keyword{datasets}

