% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_buffered.R
\docType{class}
\name{TBufferedTransportFactory}
\alias{TBufferedTransportFactory}
\title{TBufferedTransportFactory}
\format{An \code{\link{R6Class}} generator object}
\usage{
TBufferedTransportFactory
}
\description{
TBufferedTransportFactory generates TBufferedTransport.
}
\keyword{datasets}
