% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_memory.R
\docType{class}
\name{TMemoryBuffer}
\alias{TMemoryBuffer}
\title{TMemoryBuffer}
\format{An \code{\link[R6]{R6Class}} generator object}
\usage{
TMemoryBuffer
}
\description{
Wraps a raw array as a TTransport.
}
\keyword{datasets}
