% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_socket.R
\docType{class}
\name{TSocket}
\alias{TSocket}
\title{TSocket}
\format{An \code{\link{R6Class}} generator object}
\usage{
TSocket
}
\description{
Socket implementation for client side.
}
\keyword{datasets}
